[SPARK-41400][CONNECT] Remove Connect Client Catalyst Dependency
### What changes were proposed in this pull request?
This PR decouples the Spark Connect Scala Client from Catalyst, it now used SQL API module instead.

There were quite a few changes we still needed to make:
- For testing we needed a bunch of utilities. I have moved these to common-utils.
- I have moved bits and pieces of IntervalUtils to SparkIntervalUtils.
- A lot of small fixes.

### Why are the changes needed?
This reduces the client's dependency tree from ~300 MB of deps to ~30MB. This makes it easier to use the client when you are developing connect applications. On top of this the reduced dependency graph also means folks will be less affected by the clients' classpath.

### Does this PR introduce _any_ user-facing change?
Yes. It changes the classpath exposed by the Spark Connect Scala Client.

### How was this patch tested?
Existing tests.

Closes #42184 from hvanhovell/SPARK-41400-v1.

Authored-by: Herman van Hovell <herman@databricks.com>
