+++ connector/connect/client/jvm/src/main/scala/org/apache/spark/sql/connect/client/arrow/ArrowSerializer.scala
@@ -37,7 +37,7 @@ import org.apache.spark.sql.Row
 import org.apache.spark.sql.catalyst.DefinedByConstructorParams
 import org.apache.spark.sql.catalyst.encoders.AgnosticEncoder
 import org.apache.spark.sql.catalyst.encoders.AgnosticEncoders._
-import org.apache.spark.sql.catalyst.util.{IntervalUtils, SparkDateTimeUtils}
+import org.apache.spark.sql.catalyst.util.{SparkDateTimeUtils, SparkIntervalUtils}
 import org.apache.spark.sql.errors.ExecutionErrors
 import org.apache.spark.sql.types.Decimal
 import org.apache.spark.sql.util.ArrowUtils
@@ -191,11 +191,14 @@ object ArrowSerializer {
       allocator: BufferAllocator,
       timeZoneId: String): ByteString = {
     val serializer = new ArrowSerializer[T](enc, allocator, timeZoneId)
-    serializer.reset()
-    input.foreach(serializer.append)
-    val output = ByteString.newOutput()
-    serializer.writeIpcStream(output)
-    output.toByteString
+    try {
+      input.foreach(serializer.append)
+      val output = ByteString.newOutput()
+      serializer.writeIpcStream(output)
+      output.toByteString
+    } finally {
+      serializer.close()
+    }
   }
 
   /**
@@ -313,12 +316,12 @@ object ArrowSerializer {
       case (DayTimeIntervalEncoder, v: DurationVector) =>
         new FieldSerializer[Duration, DurationVector](v) {
           override def set(index: Int, value: Duration): Unit =
-            vector.setSafe(index, IntervalUtils.durationToMicros(value))
+            vector.setSafe(index, SparkIntervalUtils.durationToMicros(value))
         }
       case (YearMonthIntervalEncoder, v: IntervalYearVector) =>
         new FieldSerializer[Period, IntervalYearVector](v) {
           override def set(index: Int, value: Period): Unit =
-            vector.setSafe(index, IntervalUtils.periodToMonths(value))
+            vector.setSafe(index, SparkIntervalUtils.periodToMonths(value))
         }
       case (DateEncoder(true) | LocalDateEncoder(true), v: DateDayVector) =>
         new FieldSerializer[Any, DateDayVector](v) {
