+++ connector/connect/client/jvm/src/test/scala/org/apache/spark/sql/DataFrameNaFunctionSuite.scala
@@ -21,7 +21,7 @@ import scala.collection.JavaConverters._
 
 import org.apache.spark.SparkException
 import org.apache.spark.sql.connect.client.util.QueryTest
-import org.apache.spark.sql.internal.SQLConf
+import org.apache.spark.sql.internal.SqlApiConf
 import org.apache.spark.sql.types.{StringType, StructType}
 
 class DataFrameNaFunctionSuite extends QueryTest with SQLHelper {
@@ -388,7 +388,7 @@ class DataFrameNaFunctionSuite extends QueryTest with SQLHelper {
   }
 
   test("replace float with nan") {
-    withSQLConf(SQLConf.ANSI_ENABLED.key -> false.toString) {
+    withSQLConf(SqlApiConf.ANSI_ENABLED_KEY -> false.toString) {
       checkAnswer(
         createNaNDF().na.replace("*", Map(1.0f -> Float.NaN)),
         Row(0, 0L, 0.toShort, 0.toByte, Float.NaN, Double.NaN) ::
@@ -397,7 +397,7 @@ class DataFrameNaFunctionSuite extends QueryTest with SQLHelper {
   }
 
   test("replace double with nan") {
-    withSQLConf(SQLConf.ANSI_ENABLED.key -> false.toString) {
+    withSQLConf(SqlApiConf.ANSI_ENABLED_KEY -> false.toString) {
       checkAnswer(
         createNaNDF().na.replace("*", Map(1.0 -> Double.NaN)),
         Row(0, 0L, 0.toShort, 0.toByte, Float.NaN, Double.NaN) ::
