+++ connector/connect/client/jvm/src/test/scala/org/apache/spark/sql/PlanGenerationTestSuite.scala
@@ -44,7 +44,7 @@ import org.apache.spark.sql.functions.lit
 import org.apache.spark.sql.protobuf.{functions => pbFn}
 import org.apache.spark.sql.types._
 import org.apache.spark.unsafe.types.CalendarInterval
-import org.apache.spark.util.Utils
+import org.apache.spark.util.SparkFileUtils
 
 // scalastyle:off
 /**
@@ -131,7 +131,7 @@ class PlanGenerationTestSuite
 
   private def cleanOrphanedGoldenFile(): Unit = {
     val allTestNames = testNames.map(_.replace(' ', '_'))
-    val orphans = Utils
+    val orphans = SparkFileUtils
       .recursiveList(queryFilePath.toFile)
       .filter(g =>
         g.getAbsolutePath.endsWith(".proto.bin") ||
@@ -139,7 +139,7 @@ class PlanGenerationTestSuite
       .filter(g =>
         !allTestNames.contains(g.getName.stripSuffix(".proto.bin")) &&
           !allTestNames.contains(g.getName.stripSuffix(".json")))
-    orphans.foreach(Utils.deleteRecursively)
+    orphans.foreach(SparkFileUtils.deleteRecursively)
   }
 
   private def test(name: String)(f: => Dataset[_]): Unit = super.test(name) {
