+++ connector/connect/client/jvm/src/test/scala/org/apache/spark/sql/connect/client/util/IntegrationTestUtils.scala
@@ -23,7 +23,8 @@ import scala.util.Properties.versionNumberString
 
 import org.scalatest.Assertions.fail
 
-import org.apache.spark.util.Utils
+import org.apache.spark.SparkBuildInfo.{spark_version => SPARK_VERSION}
+import org.apache.spark.util.SparkFileUtils
 
 object IntegrationTestUtils {
 
@@ -75,14 +76,14 @@ object IntegrationTestUtils {
 
   private[sql] lazy val isSparkHiveJarAvailable: Boolean = {
     val filePath = s"$sparkHome/assembly/target/$scalaDir/jars/" +
-      s"spark-hive_$scalaVersion-${org.apache.spark.SPARK_VERSION}.jar"
+      s"spark-hive_$scalaVersion-$SPARK_VERSION.jar"
     Files.exists(Paths.get(filePath))
   }
 
   private[sql] def cleanUpHiveClassesDirIfNeeded(): Unit = {
     def delete(f: File): Unit = {
       if (f.exists()) {
-        Utils.deleteRecursively(f)
+        SparkFileUtils.deleteRecursively(f)
       }
     }
     delete(new File(s"$sparkHome/sql/hive/target/$scalaDir/classes"))
@@ -105,7 +106,7 @@ object IntegrationTestUtils {
     val jar = tryFindJar(path, sbtName, mvnName, test).getOrElse({
       val suffix = if (test) "-tests.jar" else ".jar"
       val sbtFileName = s"$sbtName(.*)$suffix"
-      val mvnFileName = s"$mvnName(.*)${org.apache.spark.SPARK_VERSION}$suffix"
+      val mvnFileName = s"$mvnName(.*)$SPARK_VERSION$suffix"
       throw new RuntimeException(s"Failed to find the jar: $sbtFileName or $mvnFileName " +
         s"inside folder: ${getTargetFilePath(path)}. This file can be generated by similar to " +
         s"the following command: build/sbt package|assembly")
@@ -136,7 +137,7 @@ object IntegrationTestUtils {
       // Maven Jar
       (f.getParent.endsWith("target") &&
         f.getName.startsWith(mvnName) &&
-        f.getName.endsWith(s"${org.apache.spark.SPARK_VERSION}$suffix"))
+        f.getName.endsWith(s"$SPARK_VERSION$suffix"))
     }
   }
