+++ core/src/main/scala/org/apache/spark/util/Utils.scala
@@ -92,7 +92,11 @@ private[spark] object CallSite {
 /**
  * Various utility methods used by Spark.
  */
-private[spark] object Utils extends Logging with SparkClassUtils {
+private[spark] object Utils
+  extends Logging
+  with SparkClassUtils
+  with SparkErrorUtils
+  with SparkFileUtils {
 
   private val sparkUncaughtExceptionHandler = new SparkUncaughtExceptionHandler
   @volatile private var cachedLocalDir: String = ""
@@ -120,16 +124,10 @@ private[spark] object Utils extends Logging with SparkClassUtils {
   })
 
   /** Serialize an object using Java serialization */
-  def serialize[T](o: T): Array[Byte] = {
-    SparkSerDeUtils.serialize(o)
-  }
+  def serialize[T](o: T): Array[Byte] = SparkSerDeUtils.serialize(o)
 
   /** Deserialize an object using Java serialization */
-  def deserialize[T](bytes: Array[Byte]): T = {
-    val bis = new ByteArrayInputStream(bytes)
-    val ois = new ObjectInputStream(bis)
-    ois.readObject.asInstanceOf[T]
-  }
+  def deserialize[T](bytes: Array[Byte]): T = SparkSerDeUtils.deserialize(bytes)
 
   /** Deserialize an object using Java serialization and the given ClassLoader */
   def deserialize[T](bytes: Array[Byte], loader: ClassLoader): T = {
@@ -254,47 +252,10 @@ private[spark] object Utils extends Logging with SparkClassUtils {
   }
 
   /**
-   * Create a directory given the abstract pathname
-   * @return true, if the directory is successfully created; otherwise, return false.
-   */
-  def createDirectory(dir: File): Boolean = {
-    try {
-      // SPARK-35907: The check was required by File.mkdirs() because it could sporadically
-      // fail silently. After switching to Files.createDirectories(), ideally, there should
-      // no longer be silent fails. But the check is kept for the safety concern. We can
-      // remove the check when we're sure that Files.createDirectories() would never fail silently.
-      Files.createDirectories(dir.toPath)
-      if ( !dir.exists() || !dir.isDirectory) {
-        logError(s"Failed to create directory " + dir)
-      }
-      dir.isDirectory
-    } catch {
-      case e: Exception =>
-        logError(s"Failed to create directory " + dir, e)
-        false
-    }
-  }
-
-  /**
-   * Create a directory inside the given parent directory. The directory is guaranteed to be
-   * newly created, and is not marked for automatic deletion.
-   */
-  def createDirectory(root: String, namePrefix: String = "spark"): File = {
-    JavaUtils.createDirectory(root, namePrefix)
-  }
-
-  /**
-   * Create a temporary directory inside the `java.io.tmpdir` prefixed with `spark`.
-   * The directory will be automatically deleted when the VM shuts down.
-   */
-  def createTempDir(): File =
-    createTempDir(System.getProperty("java.io.tmpdir"), "spark")
-
-  /**
    * Create a temporary directory inside the given parent directory. The directory will be
    * automatically deleted when the VM shuts down.
    */
-  def createTempDir(
+  override def createTempDir(
       root: String = System.getProperty("java.io.tmpdir"),
       namePrefix: String = "spark"): File = {
     val dir = createDirectory(root, namePrefix)
@@ -1177,29 +1138,13 @@ private[spark] object Utils extends Logging with SparkClassUtils {
   }
 
   /**
-   * Lists files recursively.
-   */
-  def recursiveList(f: File): Array[File] = {
-    require(f.isDirectory)
-    val result = f.listFiles.toBuffer
-    val dirList = result.filter(_.isDirectory)
-    while (dirList.nonEmpty) {
-      val curDir = dirList.remove(0)
-      val files = curDir.listFiles()
-      result ++= files
-      dirList ++= files.filter(_.isDirectory)
-    }
-    result.toArray
-  }
-
-  /**
    * Delete a file or directory and its contents recursively.
    * Don't follow directories if they are symlinks.
    * Throws an exception if deletion is unsuccessful.
    */
-  def deleteRecursively(file: File): Unit = {
+  override def deleteRecursively(file: File): Unit = {
+    super.deleteRecursively(file)
     if (file != null) {
-      JavaUtils.deleteRecursively(file)
       ShutdownHookManager.removeShutdownDeleteDir(file)
     }
   }
@@ -1444,16 +1389,6 @@ private[spark] object Utils extends Logging with SparkClassUtils {
     }
   }
 
-  /**
-   * Execute a block of code that returns a value, re-throwing any non-fatal uncaught
-   * exceptions as IOException. This is used when implementing Externalizable and Serializable's
-   * read and write methods, since Java's serializer will not report non-IOExceptions properly;
-   * see SPARK-4080 for more context.
-   */
-  def tryOrIOException[T](block: => T): T = {
-    SparkErrorUtils.tryOrIOException(block)
-  }
-
   /** Executes the given block. Log non-fatal errors if any, and only throw fatal errors */
   def tryLogNonFatalError(block: => Unit): Unit = {
     try {
@@ -1461,38 +1396,6 @@ private[spark] object Utils extends Logging with SparkClassUtils {
     } catch {
       case NonFatal(t) =>
         logError(s"Uncaught exception in thread ${Thread.currentThread().getName}", t)
-    }
-  }
-
-  /**
-   * Execute a block of code, then a finally block, but if exceptions happen in
-   * the finally block, do not suppress the original exception.
-   *
-   * This is primarily an issue with `finally { out.close() }` blocks, where
-   * close needs to be called to clean up `out`, but if an exception happened
-   * in `out.write`, it's likely `out` may be corrupted and `out.close` will
-   * fail as well. This would then suppress the original/likely more meaningful
-   * exception from the original `out.write` call.
-   */
-  def tryWithSafeFinally[T](block: => T)(finallyBlock: => Unit): T = {
-    var originalThrowable: Throwable = null
-    try {
-      block
-    } catch {
-      case t: Throwable =>
-        // Purposefully not using NonFatal, because even fatal exceptions
-        // we don't want to have our finallyBlock suppress
-        originalThrowable = t
-        throw originalThrowable
-    } finally {
-      try {
-        finallyBlock
-      } catch {
-        case t: Throwable if (originalThrowable != null && originalThrowable != t) =>
-          originalThrowable.addSuppressed(t)
-          logWarning(s"Suppressing exception in finally: ${t.getMessage}", t)
-          throw originalThrowable
-      }
     }
   }
 
@@ -2078,16 +1981,6 @@ private[spark] object Utils extends Logging with SparkClassUtils {
       case _ =>
         true
     }
-  }
-
-  /**
-   * Return a well-formed URI for the file described by a user input string.
-   *
-   * If the supplied path does not contain a scheme, or is a relative path, it will be
-   * converted into an absolute path with a file:// scheme.
-   */
-  def resolveURI(path: String): URI = {
-    SparkFileUtils.resolveURI(path)
   }
 
   /** Resolve a comma-separated list of paths. */
@@ -2757,11 +2650,6 @@ private[spark] object Utils extends Logging with SparkClassUtils {
       s"${DYN_ALLOCATION_INITIAL_EXECUTORS.key}, ${DYN_ALLOCATION_MIN_EXECUTORS.key} and " +
         s"${EXECUTOR_INSTANCES.key}")
     initialExecutors
-  }
-
-  def tryWithResource[R <: Closeable, T](createResource: => R)(f: R => T): T = {
-    val resource = createResource
-    try f.apply(resource) finally resource.close()
   }
 
   /**
